Add files via upload
Link : https://www.analystbuilder.com/questions/bad-bonuses-Eqhib
Everyone at Analyst Builder is supposed to receive a bonus at the end of the year.

Unfortunately some people didn't receive their bonus as was promised.

Write a query to determine the employees who did not receive their bonus so we can notify accounting.

Return their id and name in the output. Order the id from lowest to highest.